+++ stacker/hooks/utils.py
@@ -6,7 +6,10 @@ import sys
 import collections
 import logging
 
+from ..exceptions import FailedVariableLookup
+from ..variables import Variable, resolve_variables
 from stacker.util import load_object_from_string
+
 
 logger = logging.getLogger(__name__)
 
@@ -45,8 +48,27 @@ def handle_hooks(stage, hooks, provider, context):
     for hook in hooks:
         data_key = hook.data_key
         required = hook.required
-        kwargs = hook.args or {}
         enabled = hook.enabled
+
+        if isinstance(hook.args, dict):
+            args = [Variable(k, v) for k, v in hook.args.items()]
+            try:  # handling for output or similar being used in pre_build
+                resolve_variables(args, context, provider)
+            except FailedVariableLookup as err:
+                # pylint: disable=no-member
+                if 'pre' in stage and \
+                        "NoneType" in err.message:  # excludes detailed errors
+                    logger.error("Lookups that change the order of "
+                                 "execution, like 'output', can only be "
+                                 "used in 'post_*' hooks. Please "
+                                 "ensure that the hook being used does "
+                                 "not rely on a stack, hook_data, or "
+                                 "context that does not exist yet.")
+                raise err
+            kwargs = {v.name: v.value for v in args}
+        else:
+            kwargs = hook.args or {}
+
         if not enabled:
             logger.debug("hook with method %s is disabled, skipping",
                          hook.path)